+++ examples/tproxy/src/main.rs
@@ -68,10 +68,15 @@ fn main() -> Result<()> {
 
     // Set constants
     let mut open_object = MaybeUninit::uninit();
-    let open_skel = skel_builder.open(&mut open_object)?;
-    open_skel.maps.rodata_data.target_port = opts.port.to_be();
-    open_skel.maps.rodata_data.proxy_addr = proxy_addr.to_be();
-    open_skel.maps.rodata_data.proxy_port = opts.proxy_port.to_be();
+    let mut open_skel = skel_builder.open(&mut open_object)?;
+    let rodata = open_skel
+        .maps
+        .rodata_data
+        .as_deref_mut()
+        .expect("`rodata` is not memory mapped");
+    rodata.target_port = opts.port.to_be();
+    rodata.proxy_addr = proxy_addr.to_be();
+    rodata.proxy_port = opts.proxy_port.to_be();
 
     // Load into kernel
     let skel = open_skel.load()?;